Allow `polyphony.compiler` to be executed
This change allows for `polyphony.compiler` to be run as a script
when polyphony has been installed:

    $ python3 -m polyphony.compiler

It can often be preferable to execute installed packages in this
manner as it ensures the script is executed from the correct
path and will behave properly with virtual environments.

It is also 'correct' for the __main__ module in a package to be
runnable as a script.